# openapi.yaml
openapi: 3.0.3
info:
  title: Hiking Buddies API
  description: API for the Hiking Buddies application - connecting hikers and managing outdoor events
  version: 1.0.0
  contact:
    name: Hiking Buddies Team
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://www.hiking-buddies.com/api
    description: Production server
security:
  - cookieAuth: []
paths:
...
  /auth/login:
    post:
      summary: User login
      description: Authenticate user and create session
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: User's username
                  example: "johndoe"
                password:
                  type: string
                  description: User's password
                  format: password
                  example: "password123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  username:
                    type: string
                    example: "johndoe"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /dashboard:
    get:
      summary: Get user dashboard
      description: Retrieve current user's dashboard data (PK inferred from JWT cookie)
      tags:
        - User
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User dashboard data
          content:
            application/json:
              schema:
                type: object
                description: Dashboard content (structure depends on backend)
        '401':
          description: Not authenticated or session expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authToken
      description: JWT authentication token stored in HTTP-only cookie
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Failed to fetch events"
...